Fix task services

+++ workflow-server/src/services/task.service.ts
@@ -3,6 +3,7 @@ import { TaskRepository, UserRepository, BoardRepository } from "../repositories
 import { TaskMapper, ColumnTaskMapper } from "../mappers/index.js";
 import { TaskDTO, ColumnTaskDTO } from "../types/dto/index.js";
 import { AuthUser } from "@/types/utils.type.js";
+import { HttpError, NotFoundError } from "routing-controllers";
 
 @Service()
 export class TaskService {
@@ -28,6 +29,11 @@ export class TaskService {
   }
 
   async addTaskToColumn(boardId: string, columnId: string, taskId: string, index?: number): Promise<void> {
+    const columnsWithTasks = await this.taskRepository.getBoardTasks(boardId);
+    const columnWithTasks = columnsWithTasks.find((column) => column._id.equals(columnId));
+    if (!columnWithTasks) {
+      throw new NotFoundError("column does not exist");
+    }
     await this.taskRepository.addTaskToColumn(taskId, boardId, columnId, index);
   }
 
@@ -40,23 +46,36 @@ export class TaskService {
     await this.addTaskToColumn(boardId, columnId, taskId, rowIndex);
   }
 
-  async getAllColumnTasks(boardId: string, columnId?: string): Promise<ColumnTaskDTO | ColumnTaskDTO[]> {
+  async getAllColumnTasks(boardId: string): Promise<ColumnTaskDTO[]> {
     const columnsWithTasks = await this.taskRepository.getBoardTasks(boardId);
 
-    if (columnId) {
-      const columnWithTasks = columnsWithTasks.find((column) => column._id.equals(columnId));
-      return ColumnTaskMapper(columnWithTasks);
-    }
     return columnsWithTasks.map(ColumnTaskMapper);
+  }
+
+  async getColumnTasks(boardId: string, columnId: string): Promise<ColumnTaskDTO> {
+    const columnsWithTasks = await this.taskRepository.getBoardTasks(boardId);
+
+    const columnWithTasks = columnsWithTasks.find((column) => column._id.equals(columnId));
+    if (!columnWithTasks) {
+      throw new NotFoundError("Column does not exist");
+    }
+
+    return ColumnTaskMapper(columnWithTasks);
   }
 
   async getTask(taskId: string): Promise<TaskDTO> {
     const task = await this.taskRepository.getById(taskId);
+    if (!task) {
+      throw new NotFoundError("Task does not exist");
+    }
     return TaskMapper(task);
   }
 
   async getTaskBoardId(taskId: string): Promise<string> {
     const task = await this.taskRepository.getById(taskId);
+    if (!task) {
+      throw new NotFoundError("Task does not exist");
+    }
     return task.board.toString();
   }
 
@@ -67,23 +86,56 @@ export class TaskService {
   }
 
   async updateTask(taskId: string, taskData: any) {
-    // await this.taskRepository.
-    // no opt
+    const task = await this.taskRepository.getById(taskId);
+    if (!task) {
+      throw new NotFoundError("Task does not exist");
+    }
+    task.title = taskData.title ?? task.title;
+    task.description = taskData.description ?? task.description;
+    return await this.taskRepository.save(task);
+  }
+
+  async isMemberTaskAssignee(taskId: string, userId: string): Promise<boolean> {
+    const task = await this.taskRepository.getById(taskId);
+    const assignee = task.assignees.find(({ _id }) => _id.equals(userId));
+    return !!assignee;
   }
 
   async addAssigneeToTask(taskId: string, userId: string): Promise<void> {
+    const isAssignee = await this.isMemberTaskAssignee(taskId, userId);
+    if (isAssignee) {
+      throw new HttpError(400, "User is already assigned to this task");
+    }
     await this.taskRepository.addTaskAssignee(taskId, userId);
   }
 
   async removeAssigneeFromTask(taskId: string, userId: string): Promise<void> {
+    const isAssignee = await this.isMemberTaskAssignee(taskId, userId);
+    if (!isAssignee) {
+      throw new HttpError(400, "User is not assigned to this task");
+    }
     await this.taskRepository.removeTaskAssignee(taskId, userId);
   }
 
+  async doesTaskContainTag(taskId: string, tagsId: string): Promise<boolean> {
+    const task = await this.taskRepository.getById(taskId);
+    const foundTag = task.tags.find(({ _id }) => _id.equals(tagsId));
+    return !!foundTag;
+  }
+
   async addTagToTask(taskId: string, tagsId: string): Promise<void> {
+    const doesContainTag = await this.doesTaskContainTag(taskId, tagsId);
+    if (doesContainTag) {
+      throw new HttpError(400, "Task already contains this tag");
+    }
     await this.taskRepository.addTaskTag(taskId, tagsId);
   }
 
   async removeTagFromTask(taskId: string, tagsId: string): Promise<void> {
+    const doesContainTag = await this.doesTaskContainTag(taskId, tagsId);
+    if (!doesContainTag) {
+      throw new HttpError(400, "Task does not contains this tag");
+    }
     await this.taskRepository.removeTaskTag(taskId, tagsId);
   }
 }
